Remove batch bundle, split of reading of data from txs
