+++ op-node/rollup/derive/batch.go
@@ -20,15 +20,6 @@ import (
 //
 // An empty input is not a valid batch.
 //
-// Batch-bundle format
-// first byte is type followed by bytestring
-//
-// payload := RLP([batch_0, batch_1, ..., batch_N])
-// bundleV1 := BatchBundleV1Type ++ payload
-// bundleV2 := BatchBundleV2Type ++ compress(payload)  # TODO: compressed bundle of batches
-//
-// An empty input is not a valid bundle.
-//
 // Note: the type system is based on L1 typed transactions.
 
 // encodeBufferPool holds temporary encoder buffers for batch encoding
@@ -38,11 +29,6 @@ var encodeBufferPool = sync.Pool{
 
 const (
 	BatchV1Type = iota
-)
-
-const (
-	BatchBundleV1Type = iota
-	BatchBundleV2Type
 )
 
 type BatchV1 struct {
@@ -55,46 +41,6 @@ type BatchV1 struct {
 type BatchData struct {
 	BatchV1
 	// batches may contain additional data with new upgrades
-}
-
-func DecodeBatches(config *rollup.Config, r io.Reader) ([]*BatchData, error) {
-	var typeData [1]byte
-	if _, err := io.ReadFull(r, typeData[:]); err != nil {
-		return nil, fmt.Errorf("failed to read batch bundle type byte: %v", err)
-	}
-	switch typeData[0] {
-	case BatchBundleV1Type:
-		var out []*BatchData
-		if err := rlp.Decode(r, &out); err != nil {
-			return nil, fmt.Errorf("failed to decode v1 batches list: %v", err)
-		}
-		return out, nil
-	case BatchBundleV2Type:
-		// TODO: implement compression of a bundle of batches
-		return nil, errors.New("bundle v2 not supported yet")
-	default:
-		return nil, fmt.Errorf("unrecognized batch bundle type: %d", typeData[0])
-	}
-}
-
-func EncodeBatches(config *rollup.Config, batches []*BatchData, w io.Writer) error {
-	// default to encode as v1 (no compression). Config may change this in the future.
-	bundleType := byte(BatchBundleV1Type)
-
-	if _, err := w.Write([]byte{bundleType}); err != nil {
-		return fmt.Errorf("failed to encode batch type")
-	}
-	switch bundleType {
-	case BatchBundleV1Type:
-		if err := rlp.Encode(w, batches); err != nil {
-			return fmt.Errorf("failed to encode RLP-list payload of v1 bundle: %v", err)
-		}
-		return nil
-	case BatchBundleV2Type:
-		return errors.New("bundle v2 not supported yet")
-	default:
-		return fmt.Errorf("unrecognized batch bundle type: %d", bundleType)
-	}
 }
 
 // EncodeRLP implements rlp.Encoder
